o rearranging the begin/end download code. the Daemon now handles beginDownload()ing. There is a funky case that can cause the downloaders to beginDownload() again after the fact
I saw this occur with a pool of 4 connections. 3 connections
received Empty from the queue, then afterwards, the fourth
connection timed out. Since the others were finished, when that
fourth connection reconnected it assumed it was starting over
(beginDownload) because of the absence of the other connections.
this seems to be a general hellanzb bug (not specific to this
branch)

o When requeueing a segment (due to connectionLost), call
factory.fetchNextNZBSegment. The example I just described
benefits from this -- the fourth connection will now trigger
the other 3 connections that received Empty from the queue
to take another look at the queue (and immediately download
the requeued segment)

o added a NZBQueue.requeue function that handles normal
requeueing properly with the new RetryQueue

o added/using RetryQueue.clear

o fixed a bug in the cancel call. our connections could end
up having fetchNextNZBSegment called twice, with one of the
calls occurring when the connection is in sort of a zombie
state (immediately after cancelling). this segments to be
popped off the queue and attempted to be downloaded before
the connections have been re-made (right after being
disconnected by cancel). and you end up completely losing
the segments (they aren't ever downloaded). i think this
might be a general hellanzb bug too (and quite a nasty one)